+++ src/rgw/rgw_sal_sfs.h
@@ -400,6 +400,11 @@ class SFStore : public Store {
     }
   }
 
+  void _delete_bucket(const std::string & name) {
+    std::lock_guard l(buckets_map_lock);
+    buckets.erase(name);
+  }
+
   std::list<sfs::BucketRef> bucket_list() {
 
     _refresh_buckets_safe();